RestaurantManagementSystem-main/requirements.txt
pillow==11.2.1
beautifulsoup4==4.12.3